# libs/domains/agenstra/backend/feature-agent-controller/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Controller API
  version: 1.0.0
  description: HTTP API for managing clients and proxied agent operations
  contact:
    name: ForePath
    email: hi@forepath.io
    url: https://forepath.io
servers:
  - url: http://localhost:3100/api
    description: Local development
  - url: http://localhost:3100
    description: Root (OpenTelemetry metrics; outside /api prefix)
security:
  - bearerAuth: []
paths:
...
  /clients/{id}/agents/{agentId}/opencode/mcp:
    get:
      summary: List live MCP server statuses for an Environment
      description: >
        Environment-only. Proxies OpenCode `GET /mcp` via the agent manager. Use to detect
        servers that need interactive OAuth (`needs_auth` / `needs_client_registration`).
      operationId: listClientAgentOpencodeMcpStatuses
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: MCP server statuses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpencodeMcpStatusListResponseDto'
  /clients/{id}/agents/{agentId}/opencode/mcp/{name}/auth:
    post:
      summary: Start interactive MCP OAuth for an Environment MCP server
      description: >
        Environment-only. Builds a signed public callback URL, sets it as the worker
        oauth.redirectUri, then proxies OpenCode `POST /mcp/{name}/auth`.
      operationId: startClientAgentOpencodeMcpAuth
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OAuth flow started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpencodeMcpAuthStartResponseDto'
        '400':
          description: Agent not running, MCP OAuth unsupported, or callback env not configured
    delete:
      summary: Remove MCP OAuth credentials for an Environment MCP server
      operationId: removeClientAgentOpencodeMcpAuth
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OAuth credentials removed
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success:
                    type: boolean
                    enum: [true]
  /clients/{id}/agents/{agentId}/opencode/mcp/{name}/auth/callback:
    post:
      summary: Complete MCP OAuth with authorization code
      operationId: completeClientAgentOpencodeMcpAuth
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: name
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  minLength: 1
      responses:
        '200':
          description: Updated MCP status after OAuth
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpencodeMcpServerStatusDto'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
...
  schemas:
...
    OpencodeMcpServerStatusDto:
      type: object
      required: [name, status]
      properties:
        name:
          type: string
        status:
          type: string
          enum:
            [connected, disabled, failed, needs_auth, needs_client_registration]
        error:
          type: string
    OpencodeMcpStatusListResponseDto:
      type: object
      required: [servers]
      properties:
        servers:
          type: array
          items:
            $ref: '#/components/schemas/OpencodeMcpServerStatusDto'
    OpencodeMcpAuthStartResponseDto:
      type: object
      required: [authorizationUrl]
      properties:
        authorizationUrl:
          type: string
          format: uri
        oauthState:
          type: string
...